requirements.txt
blinker==1.9.0
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.3
Flask-WTF==1.3.0
greenlet==3.4.0
idna==3.13
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
SQLAlchemy==2.0.49
typing_extensions==4.15.0
Werkzeug==3.1.8
WTForms==3.2.1
Pillow==12.2.0

